backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.35.0
sqlmodel==0.0.22
psycopg[binary]==3.2.4
openai-agents==0.6.9
fastmcp==3.0.0b1
openai>=1.58.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
